# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.4",
        "title": "File Creation Service",
        "description": "**User Name and Password**\n* User Name: `user`\n* Password: `pass`"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "security": [
        {
            "basic_auth": []
        }
    ],
    "paths": {
        "/users": {
            "post": {
                "tags": ["Users"],
                "operationId": "createUser",
                "description": "Creates `user` object.\nDoes **require** createuser permission.\n",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    },
                    {
                        "createUserSecurity": []
                    }
                ]
            },
            "get": {
                "tags": ["Users"],
                "operationId": "getUsers",
                "description": "Gets all registered `User` objects.\n",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": ["Users"],
                "description": "Get specified `User` object",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            },
            "delete": {
                "tags": ["Users"],
                "description": "Delete specified `User` object",
                "operationId": "deleteUser",
                "responses": {
                    "200": {
                        "description": "User Deleted"
                    }
                },
                "security": [
                    {
                        "deleteUserSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            },
            "put": {
                "tags": ["Users"],
                "description": "Update specified `User` object",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "JSON User Object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User object updated"
                    }
                },
                "security": [
                    {
                        "createUserSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }
            ]
        },
        "/projects": {
            "get": {
                "tags": ["Projects"],
                "description": "Get registered `Project` objects",
                "operationId": "getProjects",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Project"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            },
            "post": {
                "tags": ["Projects"],
                "description": "Create new `Project` object",
                "operationId": "createProject",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "JSON Project Object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created `Project` Object Successful",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                },
                "security": [
                    {
                        "createProjectSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": ["Projects"],
                "description": "Get specified `Project` Object",
                "operationId": "getProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            },
            "put": {
                "tags": ["Projects"],
                "operationId": "updateProject",
                "description": "Update specified `Project` Object",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated `Project` Object Successfully",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                },
                "security": [
                    {
                        "createProjectSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            },
            "delete": {
                "tags": ["Projects"],
                "operationId": "deleteProject",
                "description": "Delete specified `Project` Object",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response"
                    }
                },
                "security": [
                    {
                        "deleteProjectSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            }
        },
        "/projects/{id}/file": {
            "post": {
                "tags": ["Projects"],
                "description": "Create file/folder in storage",
                "operationId": "createFile",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewFile"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created file/folder"
                    }
                },
                "security": [
                    {
                        "createFileSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            },
            "get": {
                "tags": ["Projects"],
                "description": "Get Info On Specified File/Folder",
                "operationId": "getFile",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewFile"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/File"
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            }
        },
...
        "/projects/{id}/user": {
            "get": {
                "tags": ["Projects", "Users"],
                "description": "Get `User` Objects Associated With Specified `Project` Object",
                "operationId": "getProjectUsers",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic_auth": []
                    }
                ]
            },
            "put": {
                "tags": ["Projects", "Users"],
                "description": "Add `User` Object to Specified `Project` Object",
                "operationId": "updateProjectUsers",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                ],
                "security": [
                    {
                        "createProjectSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }
            ]
        },
        "/acl/schema": {
            "post": {
                "tags": ["ACLSchema"],
                "description": "Create New `ACLSchema`",
                "operationId": "createACLSchema",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewACLSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created ACL Schema Successfully",
                        "schema": {
                            "$ref": "#/definitions/ACLSchema"
                        }
                    }
                },
                "security": [
                    {
                        "createACLSchemaSecurity": []
                    },
                    {
                        "basic_auth": []
                    }
                ]
            },
            "get": {
                "tags": ["ACLSchema"],
                "description": "Get Registered `ACLSchema` Objects",
                "operationId": "getACLSchemas",
                "security": [
                    {
                        "basic_auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ACLSchema"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/NewUser"
                },
                {
                    "required": [
                        "id"
                    ],
                    "properties": {
                        "id": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "created": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "modified": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                }
            ]
        },
        "NewUser": {
            "type": "object",
            "required": [
                "name",
                "password",
                "permission"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "permission": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "NewACLSchema": {
            "type": "object",
            "required": [
                "name",
                "acl"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "acl": {
                    "$ref": "#/definitions/DACL"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "basic_auth": {
            "type": "basic",
            "description": "HTTP Header Based Authentication"
        },
        "createUserSecurity": {
            "type": "basic",
            "description": "User with Ability to Create `User` Objects"
        },
        "deleteUserSecurity": {
            "type": "basic",
            "description": "User with Ability to Delete `User` Objects"
        },
        "createProjectSecurity": {
            "type": "basic",
            "description": "User with Ability to Create `Project` Objects"
        },
        "deleteProjectSecurity": {
            "type": "basic",
            "description": "User with Ability to Delete `Project` Objects"
        },
        "createFileSecurity": {
            "type": "basic",
            "description": "User with Ability to Create Files"
        },
        "setACLSecurity": {
            "type": "basic",
            "description": "User with Ability to Set ACL on Files"
        },
        "createACLSchemaSecurity": {
            "type": "basic",
            "description": "User with Ability to CRUD `ACLSchema` Objects"
        },
        "deleteACLSchemaSecurity": {
            "type": "basic",
            "description": "User wtih ability to Delete `ACLSchema` Objects"
        }
    }
}